+++ src/Components/SysTable/SysTable.js
@@ -1,20 +1,15 @@
-/* eslint-disable no-unused-vars */
 import './SysTable.scss';
 
-import React, {
-  useContext,
-  useEffect,
-  useCallback,
-  useRef,
-  useState,
-} from 'react';
+import React, { useContext, useEffect, useRef, useState } from 'react';
 
-import { useFetchSystems, buildApiFilters } from './hooks';
+import { useFetchSystems } from './hooks';
+import { buildActiveFiltersConfig, buildApiFilters } from './Helpers';
 import {
   defaultOnLoad,
   columns,
   appendDirection,
   findColumnByKey,
+  mergedColumns,
 } from './constants';
 
 import { InventoryTable } from '@redhat-cloud-services/frontend-components/Inventory';
@@ -28,7 +23,6 @@ import { RegistryContext } from '../../store';
 import { globalFilters } from '../../store/cache';
 import { useDispatch } from 'react-redux';
 import matchFilter from './matchFilter';
-import { buildActiveFiltersConfig } from './Helpers';
 import { useWorkspaceFeatureFlag } from '../../Utilities/Hooks';
 import { useIntl } from 'react-intl';
 
@@ -45,30 +39,6 @@ const SysTable = () => {
   }, [dispatch]);
   const isWorkspaceEnabled = useWorkspaceFeatureFlag();
 
-  const mergedColumns = (defaultColumns) => {
-    return columns(isWorkspaceEnabled)
-      .map((column) => {
-        const isStringCol = typeof column === 'string';
-
-        const defaultColumn = defaultColumns.find(
-          (defaultCol) => defaultCol.key === column.key
-        );
-        //this step is necessary so we could compare the default columns from column object at the top of the file
-        //with the columns that come from the inventory, so if inventory hide a column behind feature flag
-        //it will not appear when the flag is turned off
-        return column.requiresDefault && defaultColumn === undefined
-          ? undefined
-          : {
-              ...(isStringCol ? { key: column } : column),
-              ...defaultColumn,
-              props: {
-                ...defaultColumn?.props,
-                ...column?.props,
-              },
-            };
-      })
-      .filter(Boolean);
-  };
   const fetchSystems = useFetchSystems({
     query: GET_SYSTEM_TABLE,
   });
@@ -94,7 +64,7 @@ const SysTable = () => {
           operatingSystem: false,
           hostGroupFilter: false,
         }}
-        columns={mergedColumns}
+        columns={mergedColumns(isWorkspaceEnabled)}
         hasCheckbox={false}
         showTags
         ref={inventory}
