+++ src/Components/SysTable/constants.js
@@ -139,3 +139,28 @@ export const matchOptions = [
   { label: 'Matched', value: true },
   { label: 'Not Matched', value: false },
 ];
+
+export const mergedColumns = (isWorkspaceEnabled) => (defaultColumns) => {
+  return columns(isWorkspaceEnabled)
+    .map((column) => {
+      const isStringCol = typeof column === 'string';
+
+      const defaultColumn = defaultColumns.find(
+        (defaultCol) => defaultCol.key === column.key
+      );
+      //this step is necessary so we could compare the default columns from column object at the top of the file
+      //with the columns that come from the inventory, so if inventory hide a column behind feature flag
+      //it will not appear when the flag is turned off
+      return column.requiresDefault && defaultColumn === undefined
+        ? undefined
+        : {
+            ...(isStringCol ? { key: column } : column),
+            ...defaultColumn,
+            props: {
+              ...defaultColumn?.props,
+              ...column?.props,
+            },
+          };
+    })
+    .filter(Boolean);
+};
